requirements.txt
fastapi
python-multipart
pillow
uvicorn
